requirements.txt
streamlit==1.32.0
pandas
numpy
scikit-learn
sentence-transformers==2.2.2
faiss-cpu
rapidfuzz
rank-bm25
nltk